Remodel Point of Interest resource and use STI
The data for a tour is very similar to the data for a point of interest. That is why this commit will create a new superclass called attraction which will  have all the behavoiur which  POIs and Tours have in common. In the database this will be implemented with STI because this is the DRYest way. Also because Tours only have two more attributes than POIs.

* Possible risks for the future: If the subclasses get more attributes which are different, queries will get longer and more complicated because of alot of null values. This is why classis inheritance with a specific model for each resource might be the better solution. Should be discussed

#10