fix: join of `or` and `not` statements
Providing an Array in a where syntax leads to an `or`-sql statement.
But in conjunction with an `.not()` the whole statement would be negated instead of being an `or` between all single variants delivered by the array.

`WHERE NOT ((`fixed_dates`.`date_end` = NULL OR `fixed_dates`.`date_end` IS NULL))`
vs
`WHERE ((`fixed_dates`.`date_start` IS NOT NULL OR `fixed_dates`.`date_start` != NULL)`

SVA-834