feat(profile): Stay logged in after sign up
https://github.com/baschtl/devise-token_authenticatable

enable reset of the authentication token before the model is saved,
defaults to false
config.should_reset_authentication_token = false

This is the only way i found to keep the auth token after changing anything on model Member.
I also extended the validation time of the token to 1 month. After this time the Member hast to log in again.

MQGB-116