fix: remove old and duplicate records
sometimes new Resources are created, even there exist an old record to update to.
In this case we search for old resources which have the same external_id and are note linked to external_resources anymore and delete them.

We clean up old records if a new one is created or an existing one is updated.

SVA-1341