feat: Add global flag to municipalities
This commit adds a new `global` flag to the municipalities. The `global` flag indicates whether the data of a municipality can be used by other municipalities. The flag is added to the `municipalities` table in the database and is also included in the edit form and index view of municipalities.

the "default_scope", which is included by MunicipalityScope and should filter all data for a current municipality has been extended to also show data from Municipalities, which have the global flag.

SVA-1366