feat(categories): add mutation for creating categories with optional subcategories
This mutation allows the creation of one or more categories, including optional subcategories.

During execution, a transaction is used to ensure that all categories are successfully created or rolled back in case of an error.
Validation errors are collected and returned in the result upon creation.

Types::QueryTypes::CategoryType are reused as input arguments.

SVA-1125